Move subscription related classes into Subscription namespace
