+++ src/Subscriptions/Directives/BroadcastDirective.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace Nuwave\Lighthouse\Schema\Directives;
+namespace Nuwave\Lighthouse\Subscriptions\Directives;
 
 use Closure;
 use Nuwave\Lighthouse\Execution\Utils\Subscription;
+use Nuwave\Lighthouse\Schema\Directives\BaseDirective;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\FieldMiddleware;
