+++ src/Subscriptions/Directives/SubscriptionDirective.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace Nuwave\Lighthouse\Schema\Directives;
+namespace Nuwave\Lighthouse\Subscriptions\Directives;
 
+use Nuwave\Lighthouse\Schema\Directives\BaseDirective;
 use Nuwave\Lighthouse\Support\Contracts\Directive;
 
 /**
