+++ src/Subscriptions/SubscriptionResolverProvider.php
@@ -7,7 +7,7 @@ use GraphQL\Type\Definition\ResolveInfo;
 use Illuminate\Support\Str;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
 use Nuwave\Lighthouse\Schema\AST\ASTHelper;
-use Nuwave\Lighthouse\Schema\Directives\SubscriptionDirective;
+use Nuwave\Lighthouse\Subscriptions\Directives\SubscriptionDirective;
 use Nuwave\Lighthouse\Schema\Types\GraphQLSubscription;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Subscriptions\Exceptions\UnauthorizedSubscriber;
